# src/static/swagger-api.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "OriginVault's Cheqd Studio API for cheqd network",
    "version": "2.0.0",
    "description": "API service to create and manage DIDs, Verifiable Credentials, and DID-Linked Resources through OriginVault's Cheqd Studio",
    "contact": {
      "name": "OriginVault",
      "url": "https://originvault.com",
      "email": "support@originvault.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://github.com/originvault/cheqd-studio/blob/main/LICENSE"
    }
  },
...
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "name": "x-api-key",
        "in": "header"
      }
    },
    "schemas": {
...
      "Customer": {
        "type": "object",
        "properties": {
          "customer": {
            "type": "object",
            "properties": {
              "customerId": {
                "type": "string"
              },
              "name": {
                "type": "string"
              }
            }
          },
          "paymentAccount": {
            "type": "object",
            "properties": {
              "mainnet": {
                "type": "string"
              },
              "testnet": {
                "type": "string"
              }
            }
          }
        }
      },
      "QueryIdTokenResponseBody": {
        "type": "object",
        "properties": {
          "idToken": {
            "type": "string"
          }
        }
      },
...
      "InvalidRequest": {
        "description": "A problem with the input fields has occurred. Additional state information plus metadata may be available in the response body.",
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "InvalidRequest"
          }
        }
      },
...
      "UnauthorizedError": {
        "description": "Access token is missing or invalid",
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Unauthorized Error"
          }
        }
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
    "/account": {
      "get": {
        "tags": [
          "Account"
        ],
        "summary": "Fetch custodian-mode client details.",
        "description": "This endpoint returns the custodian-mode client details for authenticated users.",
        "responses": {
          "200": {
            "description": "The request was successful.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/schemas/InvalidRequest"
          },
          "401": {
            "$ref": "#/components/schemas/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/schemas/InternalError"
          }
        }
      }
    },
    "/account/idtoken": {
      "get": {
        "tags": [
          "Account"
        ],
        "summary": "Fetch IdToken.",
        "description": "This endpoint returns IdToken as JWT with list of user roles inside",
        "deprecated": true,
        "responses": {
          "200": {
            "description": "The request was successful.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryIdTokenResponseBody"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/schemas/InvalidRequest"
          },
          "401": {
            "$ref": "#/components/schemas/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/schemas/InternalError"
          }
        }
      }
    },
    "/account/create": {
      "post": {
        "tags": [
          "Account"
        ],
        "summary": "Create an client for an authenticated user.",
        "description": "This endpoint creates a client in the custodian-mode for an authenticated user",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/AccountCreateRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The request was successful.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryIdTokenResponseBody"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/schemas/InvalidRequest"
          },
          "401": {
            "$ref": "#/components/schemas/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/schemas/InternalError"
          }
        }
      }
    },
    "/trust-registry/accreditation/issue": {
      "post": {
        "tags": [
          "Trust Registry"
        ],
        "summary": "Publish a verifiable accreditation for a DID.",
        "description": "Generate and publish a Verifiable Accreditation for a subject DID as a DID Linked resource.",
        "operationId": "accredit-issue",
        "parameters": [
          {
            "in": "query",
            "name": "accreditationType",
            "description": "Select the type of accreditation to be issued.",
            "schema": {
              "type": "string",
              "enum": [
                "authorise",
                "accredit",
                "attest"
              ]
            },
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/AccreditationIssueRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccreditationIssueRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The request was successful.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Credential"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/schemas/InvalidRequest"
          },
          "401": {
            "$ref": "#/components/schemas/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/schemas/InternalError"
          }
        }
      }
    },
    "/trust-registry/accreditation/verify": {
      "post": {
        "tags": [
          "Trust Registry"
        ],
        "summary": "Verify a verifiable accreditation for a DID.",
        "description": "Generate and publish a Verifiable Accreditation for a subject DID as a DID Linked resource.",
        "operationId": "accredit-verify",
        "parameters": [
          {
            "in": "query",
            "name": "verifyStatus",
            "description": "If set to `true` the verification will also check the status of the accreditation. Requires the VC to have a `credentialStatus` property.",
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "in": "query",
            "name": "allowDeactivatedDid",
            "description": "If set to `true` allow to verify accreditation which based on deactivated DID.",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/AccreditationVerifyRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccreditationVerifyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The request was successful.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Credential"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/schemas/InvalidRequest"
          },
          "401": {
            "$ref": "#/components/schemas/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/schemas/InternalError"
          }
        }
      }
    },
    "/trust-registry/accreditation/revoke": {
      "post": {
        "tags": [
          "Trust Registry"
        ],
        "summary": "Revoke a Verifiable Accreditation.",
        "description": "This endpoint revokes a given Verifiable Accreditation. As input, it can take the didUrl as a string. The StatusList2021 resource should already be setup in the VC and `credentialStatus` property present in the VC.",
        "operationId": "accredit-revoke",
        "parameters": [
          {
            "in": "query",
            "name": "publish",
            "description": "Set whether the StatusList2021 resource should be published to the ledger or not. If set to `false`, the StatusList2021 publisher should manually publish the resource.",
            "required": true,
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/AccreditationRevokeRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccreditationRevokeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The request was successful.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevocationResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/schemas/InvalidRequest"
          },
          "401": {
            "$ref": "#/components/schemas/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/schemas/InternalError"
          }
        }
      }
    },
    "/trust-registry/accreditation/suspend": {
      "post": {
        "tags": [
          "Trust Registry"
        ],
        "summary": "Suspend a Verifiable Accreditation.",
        "description": "This endpoint suspends a given Verifiable Accreditation. As input, it can take the didUrl as a string. The StatusList2021 resource should already be setup in the VC and `credentialStatus` property present in the VC.",
        "operationId": "accredit-suspend",
        "parameters": [
          {
            "in": "query",
            "name": "publish",
            "description": "Set whether the StatusList2021 resource should be published to the ledger or not. If set to `false`, the StatusList2021 publisher should manually publish the resource.",
            "required": true,
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/AccreditationRevokeRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccreditationRevokeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The request was successful.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevocationResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/schemas/InvalidRequest"
          },
          "401": {
            "$ref": "#/components/schemas/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/schemas/InternalError"
          }
        }
      }
    },
    "/trust-registry/accreditation/reinstate": {
      "post": {
        "tags": [
          "Trust Registry"
        ],
        "summary": "Reinstate a Verifiable Accreditation.",
        "description": "This endpoint reinstates a given Verifiable Accreditation. As input, it can take the didUrl as a string. The StatusList2021 resource should already be setup in the VC and `credentialStatus` property present in the VC.",
        "operationId": "accredit-reinstate",
        "parameters": [
          {
            "in": "query",
            "name": "publish",
            "description": "Set whether the StatusList2021 resource should be published to the ledger or not. If set to `false`, the StatusList2021 publisher should manually publish the resource.",
            "required": true,
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/AccreditationRevokeRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccreditationRevokeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The request was successful.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevocationResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/schemas/InvalidRequest"
          },
          "401": {
            "$ref": "#/components/schemas/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/schemas/InternalError"
          }
        }
      }
    },
...
  }
}